Move anim files to mapping core
- Move map_image.cpp, map_actor.cpp, map_animations.h, map_animations.cpp to mapping core
