+++ src/chkdraft/mapping/clipboard_animations.h
@@ -1,9 +1,9 @@
 #ifndef CLIPBOARDANIMATIONS_H
 #define CLIPBOARDANIMATIONS_H
 #include <mapping_core/map_actor.h>
+#include <mapping_core/render/map_animations.h>
 #include <mapping_core/sc.h>
 #include <vector>
-#include "map_animations.h"
 
 class Clipboard;
