+++ src/mapping_core/render/map_actor.cpp
