+++ src/mapping_core/render/map_animations.cpp
