+++ src/mapping_core/render/map_image.cpp
